requirements.txt
pandas
numpy
sentence_transformers
faiss-cpu
chromadb
openai
streamlit
python-dotenv
transformers
google-genai
langchain-google-genai
langchain
gradio
google.generativeai